LogPlayer: extract the $monitorLoggerStep value from the timestamps
Previously it was hardcoded to the default value 0.2, not it should work with arbitrary step sizes.
The stepSize is now also used to give a better estimation of the total number of frames.

PlayerControls: removed the tick painting, it wasn't very meaningful due to a variable amount of "non-PlayOn frames" at the start of a log.