Make the sizing of windows more portable
This is done using preferredSize for the actual content.

This commit also fixes the issue that windows are too small on macOS when using the darklaf library. (That was actually an issue with darklaf that will probably be fixed with the next released darklaf version. Still, this approch should be more portable across different operating systems.)