Assign default team color based on team name
Instead of assigning a color always to the left team, the left team will get the left color if its team name is less in the alphabetical sorting of the two team names. Otherwise, the right color will be assigned to the left team. If a team name is specified in the config file, this will still override this behavior. Also, if only one team is present, the left team will always get the default color for the left team (if no team name is specified).