Prevent a NullPointerException in Logfile
When stopLogPlayer is called, a running log analyzer may still be
running and will be aborted.
It may take a moment for the thread to stop. At that point logfile may
already be set to null. The LogAnalyzerThread, however, called the
"finished" callback either way. In that callback, we're trying to access
logfile (which is now null) and there's our NullPointerException.

This solution does not run any callbacks once the analyzer was aborted.
Should the abortion happen just while the callback is running, we'll use
a local reference to logfile to prevent a very unlucky
NullPointerException at that point.