Improve the roboviz shell script
Instead of parsing the arguments and replacing the paths to the log files
with absolute paths, we'll just keep the current working directory. We then
refer to the Jar file through the already existing DIR variable.

This also supports setting the Dock icon on macOS if the application is
running inside of a bundled application.

And on top of that, the script should now be compatible with other
POSIX-compatible shells (not just bash).