Support running without a present config file
Until now, RoboViz would not start without a config file being present.
Now, we'll just use the default hardcoded configuration values in that
case. When attempting to save the configuration using the built-in
configuration editor, RoboViz will ask if a new configuration file
should be initialized (if none is present so far).